Moved file and removed placeholder
File moved to components/event
Removed WhyAttend placeholder
