+++ src/components/event/EventHeader.jsx
@@ -15,7 +15,7 @@ const EventHeader = () => {
         Music Makes Futures (MMF) invites you to an evening filled with
         elegance, inspiration, and purpose. Mark your calendars for August 25,
         2024, as we host our annual Back-to-School Fundraiser Gala at the
-        prestigious Le Jardin Étoilé, Route Royale, Antananarivo, Madagascar.
+        prestigious Le Jardin Étoilé, Route Royale, Antananarivo, Madagascar.
         This gala isn't just an event; it's a celebration of hope, a testament
         to the power of education, and a night that promises to make a
         significant difference in the lives of children in Madagascar.
